% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argument}
\alias{argument}
\title{Specify command line argument values}
\usage{
argument(default_val = NULL, help_string = NULL)
}
\arguments{
\item{default_val}{(number | string) (default NULL)}

\item{help_string}{(string) (default NULL)}
}
\value{
(list) with correct key value pairs for parsing
}
\description{
Specify command line argument values
}
