% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{best_fold}
\alias{best_fold}
\title{Given results from one k-fold cross-validation, identify the
fold with the best model performance.}
\usage{
best_fold(cv_results, metric = "pearson")
}
\arguments{
\item{cv_results}{(list)
A list with results from one k-fold cross-validations, as output by 
kfold_cv().}

\item{metric}{(character)
(Default 'pearson') The desired statistic to use to identify the best 
model fit. The 'best' fold is defined as that whose performance 
maximizes this metric.}
}
\value{
The integer value identifying the 'test' list element (i.e., the 
     fitted fold) with the best model performance, per the desired 
     performance metric.
}
\description{
Given results from one k-fold cross-validation, identify the
fold with the best model performance.
}
