% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{best_kfold_mod}
\alias{best_kfold_mod}
\title{Given results from multiple k-fold cross-validations, identify the
model with the best mean cross-validation performance.}
\usage{
best_kfold_mod(crossval_list)
}
\arguments{
\item{crossval_list}{(list)
A list as produced by kfold_cv(), or a list with results from multiple 
k-fold cross-validations, in which each top-level element was produced
by kfold_cv().}
}
\value{
The integer value identifying the list element with the best
     joint cross-validation performance: the best pearson correlation between
     all observations and predictions (results from all folds combined).
}
\description{
Given results from multiple k-fold cross-validations, identify the
model with the best mean cross-validation performance.
}
