% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_gof}
\alias{compute_gof}
\title{Estimate goodness-of-fit statistics for a set of predictions}
\usage{
compute_gof(observed, predicted)
}
\arguments{
\item{observed}{(numeric) observed trait values}

\item{predicted}{(numeric) predictions, aligned with `observed`}
}
\value{
named list of r_sq, pearson_corr, spearman_corr
}
\description{
Estimate goodness-of-fit statistics for a set of predictions
}
