% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_r_sq}
\alias{compute_r_sq}
\title{Compute the coefficient of determination (\eqn{R^2})}
\usage{
compute_r_sq(true_vals, predictions)
}
\arguments{
\item{true_vals}{((n,) array or vector)}

\item{predictions}{((n,) array or vector)}
}
\value{
(double)
}
\description{
Compute the coefficient of determination (\eqn{R^2})
}
