% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_bpar}
\alias{convert_bpar}
\title{Store model parameters from a bpar file.}
\usage{
convert_bpar(bpar_file)
}
\arguments{
\item{bpar_file}{(character)
The path to the bpar file with desired parameter values.}
}
\value{
A list with model parameters, BLUPs, and performance statistics, 
     as output by fit().
}
\description{
Reads a bpar file and stores model parameters as a rattaca model object, 
formatted as output by fit().
}
