% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_manual_assignment_files}
\alias{create_manual_assignment_files}
\title{Create files to use in manual RATTACA assignment}
\usage{
create_manual_assignment_files(predictions, colony_df, assign_traits, out_dir)
}
\arguments{
\item{predictions}{(character or dataframe)
R dataframe or path to the predictions file.}

\item{colony_df}{(character or dataframe)
R dataframe or path to the colony dataframe.}

\item{assign_traits}{(character)
Path to a text file listing assignment traits to include in the output file.}

\item{out_dir}{(character)
Directory path in which to save manual assignment files.}
}
\value{
None. Saves a text file of variant IDs and a csv file of training rfids 
and trait data into the same directory as the parameter file.
}
\description{
Merges a merged predictions file and colony dataframe, outputs sex-specific
csv files with rat metadata and prediction ranks for assignment traits.
}
