% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_trait_dict}
\alias{create_trait_dict}
\title{Create a RATTACA trait dictionary from a setup file}
\usage{
create_trait_dict(setup, outfile, return_dict = FALSE)
}
\arguments{
\item{setup}{(character or dataframe)
R dataframe or path to a csv dict setup file. The setup df should have
one row per trait, and the following columns: trait, project_name, 
data_file, data_var, data_dict, dict_var, h2_file, h2_var}

\item{outfile}{(character)
Path to the desired output file, a trait dictionary in csv format}

\item{return_dict}{(boolean)
(Default FALSE) Whether to return the dictionary to the R environment}
}
\value{
None. Saves a text file of variant IDs and a csv file of training rfids 
and trait data into the same directory as the parameter file.
}
\description{
Reads a trait dict setup file to extract variable names, filepaths,
and heritabilities from various input files, then formats and saves them to
a RATTACA trait dictionary
}
