% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{density_assignments}
\alias{density_assignments}
\title{Produce an assignment density plot: trait z-score density.}
\usage{
density_assignments(
  df,
  trait,
  assignment_col,
  gen = NULL,
  trait_name = NULL,
  outdir = NULL
)
}
\arguments{
\item{df}{(dataframe)
A dataframe of trait predictions for a trait of interest (as produced 
by get_ranks_zscores()), plus a column of group assignments for that 
trait (as produced by trait_groups()).}

\item{trait}{(string)
The name of the trait of interest. Must be the column header for the 
trait predictions to be plotted.}

\item{assignment_col}{(string)
The column header for the column of group assignments to be plotted.}

\item{gen}{(int)
The RATTACA generation being plotted.}

\item{trait_name}{(string)
(Default NULL) The desired trait name to use in the figure title and 
file name. Use this option to simplify naming when a trait has a long 
or unintuitive variable name.}

\item{outdir}{(string)
(Defualt NULL) The directory path in which to save the figure. If NULL, 
the figure will plot to the R console.}
}
\value{
Plots to the R console if outdir=NULL. Saves a png file if outdir is
         not NULL.
}
\description{
Reads a dataframe of predictions, z-scores, and group assignments to plot the
density of prediction z-scores, with a rug of individual z-scores colored by 
assigned high/low group.
}
