% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit}
\alias{fit}
\title{Fit marker based rrBLUP to data}
\usage{
fit(trait, genotypes)
}
\arguments{
\item{trait}{((n animals,) double vector) of trait
values.}

\item{genotypes}{((n animals, q markers) double matrix)
of animal genotypes.  Markers are assume to take
values (1, 0, -1, NA).  NA values for marker \eqn{j} are 
assigned the mean genotype over all animals \eqn{i} for
for which marker \eqn{i \neq j} and is not NA.}
}
\value{
out (list): contains elements generated by
    rrBLUP::mixed.solve and goodness-of-fit values computed
    by this function.
\itemize{
     \item{$u}
     {
         ((q markers,) array) BLUPs, marker effect
         sizes from rrBLUP::mixed.solve
     }
     \item{$u.SE}
     {
         ((q markers,) array) standard error of 
         BLUPs produced by rrBLUP::mixed.solve
     }
     \item{$beta}
     {
         (double) intercept (BLUE) inferred by 
         rrBLUP::mixed.solve
     }
     \item{$beta.SE}
     {
         (double) standard error of intercept
         produced by rrBLUP::mixed.solve
     }
     \item{$LL}
     {
         (double) negative log-likelihood produced
         by rrBLUP::mixed.solve
     }
     \item{$Ve}
     {
         (double) variance of the error term
         produced by rrBLUP::mixed.solve
     }
     \item{$Vu}
     {
         (double) variance contribution of a single
         random effects term produced by rrBLUP::mixed.solve
     }
     \item{$r_sq}{(double) coefficient of determination}
     \item{$pearson_corr}{(double)}
     \item{$spearman_corr}{(double)}
}
}
\description{
Validate input data, fit LMM to data with
rrBLUP::mixed.solve and, and compute goodness-of-fit
statistics.
}
