% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ids_file}
\alias{format_ids_file}
\title{Save a Plink-formatted file listing all samples input in a vector.}
\usage{
format_ids_file(ids, output_dir, filename = "plink_ids")
}
\arguments{
\item{ids}{(character)
A vector of sample IDs.}

\item{output_dir}{(string)
The directory in which the Plink-formatted IDs file will be 
saved.}

\item{filename}{(string)
(default 'plink_ids') The desired file name}
}
\value{
A list containing (1) the file path to the Plink-formatted 
     IDs file and (2) a vector of IDs.
}
\description{
Save a Plink-formatted file listing all samples input in a vector.
}
