% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_obs_pred}
\alias{format_obs_pred}
\title{Format trait observations and predictions into a dataframe}
\usage{
format_obs_pred(lst)
}
\arguments{
\item{lst}{(list)
A list with elements $obs (named trait observations) and $pred (trait
predictions in the same order as $obs). Each element may itself be a 
list of length k sets of observations or predictions, such as output
by kfold_cv()}
}
\value{
A dataframe of all IDs, observations, and predictions, concatenated
     across all k folds
}
\description{
Format trait observations and predictions into a dataframe
}
