% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_common_snpset}
\alias{get_common_snpset}
\title{Identify the set of SNP variants common to both the training
and test datasets}
\usage{
get_common_snpset(train_genotypes, test_genotypes)
}
\arguments{
\item{train_genotypes}{(string, list, or matrix)
Alternatively, the file path/prefix for the training Plink dataset,
a training genotype data object as produced by make_plink_dataset(), 
or a named training genotype matrix with RFID rows and variant columns}

\item{test_genotypes}{(string, list, or matrix)
Alternatively, the file path/prefix for the test Plink dataset,
a test genotype data object as produced by make_plink_dataset(), 
or a named test genotype matrix with RFID rows and variant columns}
}
\value{
A vector of all SNP variants found in both datasets.
}
\description{
Identify the set of SNP variants common to both the training
and test datasets
}
