% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_group_means}
\alias{get_group_means}
\title{Estimate per-group means for trait simulations produced during power analysis}
\usage{
get_group_means(pwr_list)
}
\arguments{
\item{pwr_list}{(list)
A list of power_analysis objects}
}
\value{
A dataframe of simulated group means for each group x replicate 
     provided in the input
}
\description{
Estimate per-group means for trait simulations produced during power analysis
}
