% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_phenotyped_ids}
\alias{get_phenotyped_ids}
\title{Save a Plink-formatted file listing all samples phenotyped
for one or all traits in a dataset.}
\usage{
get_phenotyped_ids(phenotype_data, id_column, output_dir, trait_column = NULL)
}
\arguments{
\item{phenotype_data}{(string)
File path to a csv file containing one named column of 
sample IDs and one or more named columns with phenotype
data.}

\item{id_column}{(string)
Text string with the header for the column containing
sample IDs.}

\item{output_dir}{(string)
The directory in which the Plink-formatted IDs file will be 
saved.}

\item{trait_column}{(string)
(default NULL) Text string with the header for the desired
phenotype, if only processing one. If NULL, all non-ID
columns in the file will be processed.}
}
\value{
A list containing (1) the file path to the Plink-formatted 
     IDs file and (2) a vector of phenotyped IDs.
}
\description{
Save a Plink-formatted file listing all samples phenotyped
for one or all traits in a dataset.
}
