% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ranks_zscores}
\alias{get_ranks_zscores}
\title{Given a set of trait predictions, get the rank order and the z-score
of each}
\usage{
get_ranks_zscores(predictions, trait, output_dir = NULL)
}
\arguments{
\item{predictions}{(numeric)
A named vector of trait predictions}

\item{trait}{(character)
     The name of the trait to be analyzed

#' @param output_dir (character)
     (default NULL) The directory in which to save a dataframe of 
     trait predictions, prediction ranks, and z-scores, if desired}
}
\value{
A list containing (1) the trait analyzed, (2) all prediction ranks,
     and (3) all prediction z-scores
}
\description{
Given a set of trait predictions, get the rank order and the z-score
of each
}
