% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_snps_from_bpar}
\alias{get_snps_from_bpar}
\title{Get variant IDs for all SNPs in a RATTACA bpar file}
\usage{
get_snps_from_bpar(bpar_file, outdir = NULL)
}
\arguments{
\item{bpar_file}{(string)
The file path to a bpar file}

\item{outdir}{(string)
(default NULL) The directory path in which to save a Plink-formatted 
list of variant IDs}
}
\value{
A vector of all SNP variant IDs found in the bpar file.
}
\description{
Get variant IDs for all SNPs in a RATTACA bpar file
}
