% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gof_spread}
\alias{gof_spread}
\title{Get the 'spread' of goodness-of-fit statistics for all cross validation folds}
\usage{
gof_spread(cv, outdir = NULL)
}
\arguments{
\item{cv}{(list) 
Results from a single k-fold cross validation, as output by kfold_cv()}

\item{outdir}{(character)
(default NULL) The directory in which to save the output df}
}
\value{
A dataframe of per-fold and joint goodness-of-fit statistics.
}
\description{
Extracts goodness-of-fit statistics from each fold of a cross validation, 
and estimates joint statistics across all folds
}
