% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_genotype}
\alias{is_genotype}
\title{Test whether a matrix consists of genotype data}
\usage{
is_genotype(x, imputed_vals = TRUE)
}
\arguments{
\item{x}{((n, m) array | vector | matrix) of genotype
values}

\item{imputed_vals}{(boolean) whether input contains
imputed values}
}
\value{
(boolean)
}
\description{
Valid genotype matrices should contain values in the
set \eqn{\{1,0,-1\}} if no imputation performed.  Otherwise,
genotype values \eqn{g_{ij} \in [-1, 1]}.
}
