% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kfold_cv}
\alias{kfold_cv}
\title{Perform a k-fold cross-validation on rrBLUP predictions for one trait}
\usage{
kfold_cv(data, num_folds, out_dir = NULL)
}
\arguments{
\item{data}{(list)
An aligned genotype/phenotype dataset, as produced by rattaca::align:
A list with elements $trait (character string naming the trait to
analyze), $pheno (a named vector of phenotype observations), and $geno
(a genotype matrix aligned with $pheno)}

\item{num_folds}{(int)
The desired number of folds (k) used for k-fold cross-validation}

\item{out_dir}{(character)
The path to the directory in which to save output files summarizing
cross-validation performance}
}
\value{
A list of (1) the trait name, (2) the lists of RFIDs including in the
     train and test sets of each fold, (3) the list of k model fits onto each
     training sample (as produced by fit()), and (3) the list of k model
     validations of each test sample (as produced by validate_test_preds())
}
\description{
Perform a k-fold cross-validation on rrBLUP predictions for one trait
}
