% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_existing_plink_dataset}
\alias{load_existing_plink_dataset}
\title{Produce a rattaca geno data object from an existing Plink genotype dataset.}
\usage{
load_existing_plink_dataset(file_stem)
}
\arguments{
\item{file_stem}{(string)
The file path/prefix for the Plink dataset.}
}
\value{
A rattaca genotype dataset: a list of (1) the file stem for the 
     dataset and (2) the genotype matrix.
}
\description{
Produce a rattaca geno data object from an existing Plink genotype dataset.
}
