% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_plink_dataset}
\alias{make_plink_dataset}
\title{Modify a Plink genotype dataset to produce a new Plink dataset
of system files, with option to read data into R}
\usage{
make_plink_dataset(
  input_genotypes,
  output_dir,
  outfile_prefix = "",
  samples_to_keep = NULL,
  snps_to_keep = NULL,
  maf_cutoff = NULL,
  missing_cutoff = NULL,
  hwe_cutoff = NULL,
  ld_prune = FALSE,
  ld_r2 = NULL,
  ld_window_size = NULL,
  ld_step_size = NULL,
  snp_directory = NULL,
  return_data = TRUE
)
}
\arguments{
\item{input_genotypes}{(string)
prefix of plink bim, bed, and fam files that will be
loaded and modified}

\item{output_dir}{(string)
Output directory in which to save the new Plink dataset}

\item{outfile_prefix}{(string) 
(default '') The base file name for new Plink files.
This will be appended with modifications to the original 
dataset.}

\item{samples_to_keep}{(string) 
(default NULL) If subsetting samples from the original
dataset, the path to the file listing all samples desired
for the new dataset. Must be in Plink .fam format, with 
one row per sample and two columns: one column of family IDs 
(or zeros), and one column of individual IDs, as produced
by get_phenotyped_ids(). Using this option calls the Plink
option '--keep'.}

\item{snps_to_keep}{(string) 
(default NULL) If subsetting SNP variants from the original
dataset, the path to the file listing all SNP variants desired
for the new dataset. Must be formatted with one column of 
variant IDs, one row per variant, as produced by either 
sample_snps() or sample_snps_from_plink_files(). Using this 
option calls the Plink option '--extract'.}

\item{maf_cutoff}{(double) 
(default NULL) The desired minor allele frequency cutoff. All 
variants with allele frequency in the original dataset below 
the provided threshold will be excluded from the new dataset.
Using this option calls the Plink option '--maf'.}

\item{missing_cutoff}{(double) 
(default NULL) The desired missing call rate cutoff. All variants 
with missing call rates exceeding the provided threshold in the 
original dataset will be excluded from the new dataset. Using 
this option calls the Plink option '--geno'.}

\item{hwe_cutoff}{(double) 
(default NULL) The desired cutoff for Hardy-Weinberg exact test 
P-values. All variants in the original dataset whose P-values 
following a Hardy-Weinberg equilibrium exact test fall below the 
provided threshold will be excluded from the new dataset. Using
this option calls the Plink option '--hwe'.}

\item{ld_prune}{(bool) 
(default FALSE) Whether or not to prune variants from the original
dataset based on linkage disequilibrium with other variants. Using
this option calls the Plink option '--indep-pairwise'.}

\item{ld_r2}{(double) 
(default NULL) If LD-pruning, The desired pairwise r^2 threshold.
For all variant pairs within a desired window size, if the r^2
of the pairwise correlation between variant allele counts exceeds
this threshold, only one variant will be kept in the new dataset.}

\item{ld_window_size}{(integer) 
(default NULL) If LD-pruning, the desired window size (in variant 
count) in which to estimate pairwise linkage disequilibrium.}

\item{ld_step_size}{(integer) 
(default NULL) If LD-pruning, the desired step size (in variant 
count) by which the pruning window will slide before estimating 
pairwise LD again.}

\item{snp_directory}{(string)
(default NULL) If LD-pruning, the desired directory path in which to
save pruned.in and pruned.out files}

\item{return_data}{(bool) 
(default TRUE) Whether or not to return the new dataset in the R 
environment. If TRUE, returns a list containing (1) the file 
path to the new dataset and (2) the new genotype matrix. If 
FALSE, returns only the file directory. This option is preferable 
when producing large datasets that become computationally intractable 
to process within R.}
}
\value{
A list containing (1) the file path to the new dataset and 
     (2) the new genotype matrix.
}
\description{
Modify a Plink genotype dataset to produce a new Plink dataset
of system files, with option to read data into R
}
