% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean_imputation}
\alias{mean_imputation}
\title{Mean imputation of NA values in a vector}
\usage{
mean_imputation(v)
}
\arguments{
\item{v}{((n,) vector)}
}
\value{
((n,) vector) in which NA values are replaced by
     imputed values.
}
\description{
Consider an \eqn{n} length vector \eqn{v}. If not all elements
of \eqn{v} are not NA, but a subset are.  Replace each NA
element by the mean of the elements of \eqn{v} that are not NA.
}
