% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_preds}
\alias{merge_preds}
\title{Given a directory containing multiple csv files of predictions for 
different traits, merge all files into a single csv of results for
all traits}
\usage{
merge_preds(
  directory,
  traits,
  composite_traits = NULL,
  output_dir = NULL,
  basename = NULL
)
}
\arguments{
\item{directory}{(character)
The directory path housing prediction results for multiple traits.
Within this directory, each individual trait's results must be 
inside a separate sub-directory identified by the trait name, with
a csv file named as <trait_name>_predictions.csv (as output by 
get_ranks_zscores()), e.g.: 
<directory>/<trait_name>/<trait_name>_predicitons.csv}

\item{traits}{(list or character)
Either a list or vector of trait names whose respective trait
predictions will be merged}

\item{composite_traits}{(character)
(default NULL) A vector of composite trait names whose respective 
composite trait scores will be merged}

\item{output_dir}{(character)
(default NULL) The directory in which the csv of merged trait
predictions will be saved, if desired.}

\item{basename}{(character)
(default NULL) The stem of the filename used for the merged csv
file, if desired. Final results will be saved to
<output_dir>/<basename>_merged_predictions.csv}
}
\value{
The dataframe of all merged trait predictions
}
\description{
Given a directory containing multiple csv files of predictions for 
different traits, merge all files into a single csv of results for
all traits
}
