% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mod_gof}
\alias{mod_gof}
\title{Append goodness-of-fit statistics to a model object}
\usage{
mod_gof(mod, cv)
}
\arguments{
\item{mod}{(list) 
A RATTACA model object, as output by fit()}

\item{cv}{(list) 
Results from a single k-fold cross validation, as output by kfold_cv()}
}
\value{
The input object 'mod', with goodness-of-fit statistics appendend to
     the list object.
}
\description{
Calculates joint goodness-of-fit statistics for a model fit by estimating
on all observations and predictions (appended across all folds of a 
cross validation)
}
