% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pca_plink_genotypes}
\alias{pca_plink_genotypes}
\title{Conduct a PCA on train- and test-set genotypes using Plink}
\usage{
pca_plink_genotypes(train_genotypes, test_genotypes, trait)
}
\arguments{
\item{train_genotypes}{(character)
The file path/prefix for the training Plink dataset.}

\item{test_genotypes}{(character)
The file path/prefix for the test Plink dataset.}

\item{trait}{(character)
The name of the trait to be analyzed}
}
\value{
A list of (1) the trait name, (2) the path to the Plink file
     of PCA results
}
\description{
Conduct a PCA on train- and test-set genotypes using Plink
}
