% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_assignments}
\alias{plot_assignments}
\title{Produce an assignment 'S' plot: trait rank vs. trait prediction.}
\usage{
plot_assignments(
  df,
  trait,
  assignment_col,
  jitter = c(80, 40),
  gen,
  random_seed = 1,
  trait_name = NULL,
  outdir = NULL
)
}
\arguments{
\item{df}{(dataframe)
A dataframe of trait predictions for a trait of interest (as produced 
by get_ranks_zscores()), plus a column of group assignments for that 
trait (as produced by trait_groups()).}

\item{trait}{(string)
The name of the trait of interest. Must be the column header for the 
trait predictions to be plotted.}

\item{assignment_col}{(string)
The column header for the column of group assignments to be plotted.}

\item{jitter}{(numeric)
(Default c(80,40)) A vector of jitter magnitudes for non-assigned and 
assigned points, respectively. This may need tweaking for aesthetic
plotting.}

\item{gen}{(int)
The RATTACA generation being plotted.}

\item{random_seed}{(int)
(Default 1) An integer value with which to set the seed for random
jittering. This allows consistent placement of jittered points and may 
need tweaking for aesthetic plotting.}

\item{trait_name}{(string)
(Default NULL) The desired trait name to use in the figure title and 
file name. Use this option to simplify naming when a trait has a long 
or unintuitive variable name.}

\item{outdir}{(string)
(Defualt NULL) The directory path in which to save the figure. If NULL, 
the figure will plot to the R console.}
}
\value{
Plots to the R console if outdir=NULL. Saves a png file if outdir is
         not NULL.
}
\description{
Reads a dataframe of predictions, ranks, and group assignments to plot 
prediction rank against predicted values, colored by assigned high/low group.
}
