% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_kfold}
\alias{plot_kfold}
\title{Plot the results of a k-fold cross validation to files}
\usage{
plot_kfold(kfold_results, output_dir)
}
\arguments{
\item{kfold_results}{(list)
A list with results from a k-fold cross-validation, as produced
by kfold_cv(). Must contain elements $trait (the trait name) and
$test (a list of k sets oftrait observations, predictions, and 
performance metrics from a k-fold cross-validation)}

\item{output_dir}{(string)
The directory in which the png files will be saved}
}
\value{
None. Plots are saved to files, without output into the R 
     environment
}
\description{
Plot the results of a k-fold cross validation to files
}
