% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot the results of a Plink PCA to png files}
\usage{
plot_pca(plink_pca, train_genotypes, test_genotypes, output_dir)
}
\arguments{
\item{plink_pca}{(list)
A list containing elements $trait (the trait to analyze) and $pca_file
(the path to Plink .eigenvec PCA results), as output by 
pca_plink_genotypes()}

\item{train_genotypes}{(character)
     A genotype matrix from the training set

#' @param test_genotypes (character)
     A genotype matrix from the test set

#' @param output_dir (character)
     The directory in which to save png plots}
}
\value{
None. Plots are saved to files, without output into the R 
     environment
}
\description{
Plot the results of a Plink PCA to png files
}
