% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{power_analysis}
\alias{power_analysis}
\title{Power analysis given genotypes and model simulator}
\usage{
power_analysis(
  geno_low,
  geno_high,
  sim,
  significance_level = 0.05,
  m_power_reps = 100
)
}
\arguments{
\item{geno_low}{((n_low samples, q markers) array | matrix) 
genotypes of samples whose mean predicted value is
less than geno_high}

\item{geno_high}{((n_high sample, q marker) array | matrix)
genotypes of samples whose mean predicted value is 
larger than geno_low}

\item{sim}{(function) a function from simulator closure.}

\item{significance_level}{(double)}

\item{m_power_reps}{(int)}
}
\value{
(double) statistical power
}
\description{
Power analysis given genotypes and model simulator
}
