% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict_lmm}
\alias{predict_lmm}
\title{Predict trait values from genotype}
\usage{
predict_lmm(genotypes, u, intercept)
}
\arguments{
\item{genotypes}{((n sample, q markers) array | matrix)
sample genotypes encoded on the interval [-1, 1].  the
elements that are not integers (i.e. (-1,0,1)) are assumed
to have been appropriately impurted}

\item{u}{((q markers,) vector | (q markers,1) matrix)
estimated random effects parameters}

\item{intercept}{(double)}
}
\value{
((n sample,) vector)
     trait predictions predictions
}
\description{
Compute the quantitative trait from an animal's genotype and
provided parameters. Parameters are inferred from a reference
data set.
}
