% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_trait_data}
\alias{prep_trait_data}
\title{load trait data from csv and prepare}
\usage{
prep_trait_data(
  filename,
  id_column,
  trait_name,
  format = c("numeric", "data.frame")
)
}
\arguments{
\item{filename}{(character)
path to trait csv table}

\item{id_column}{(character)
name of column with animal IDs whose values will be the data index names}

\item{trait_name}{(character)
name of column whose values consists of noramlized
trait measurements}

\item{format}{(string)
The data type in which to output results 
(either 'data.frame' or 'numeric')}
}
\value{
A dataframe with n_sample rows x 1 column, rows named with sample IDs, 
     or a numeric vector named with sample IDs.
}
\description{
Data are loaded assuming a header exists.  Using this 
     header, the row names are set to the sample id's and
     and samples with trait value equal to NA are removed.
}
