% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printout}
\alias{printout}
\title{Print time-stamped messages to stdout.}
\usage{
printout(str)
}
\arguments{
\item{str}{(character)
A text string to print.}
}
\description{
Print time-stamped messages to stdout.
}
