% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rattaca_power}
\alias{rattaca_power}
\title{Conduct a RATTACA power analysis}
\usage{
rattaca_power(
  geno_low,
  geno_high,
  sim,
  alpha = 0.05,
  tests_per_rep,
  reps,
  trait,
  outdir = NULL
)
}
\arguments{
\item{geno_low}{((n_low samples, q markers) array | matrix) 
Genotypes of samples whose mean predicted value is
less than geno_high}

\item{geno_high}{((n_high sample, q marker) array | matrix)
Genotypes of samples whose mean predicted value is 
larger than geno_low}

\item{sim}{(function) 
A function from simulator closure, as output by gen_trait_sim_closure().}

\item{tests_per_rep}{(int)
The number of bootstrap replicates to be conducted by each
experimental replicate power analysis.}

\item{reps}{(int)
The number of experimental replicates to conduct}

\item{trait}{(character)
Name of the trait being analyzed}

\item{outdir}{(character)
(default NULL) The directory path in which to save results, if provided.}

\item{significance_level}{(double)
(default 0.05) The desired alpha level at which to test differences
between groups.}
}
\value{
(dataframe) A dataframe of metadata and statistical power for all
     experimental replicates of the analysis.
}
\description{
This is a wrapper for the basic power_analysis() function. Can conduct 
multiple experimental replicates of a power analysis (each with a given 
number of bootstrap replicates) and enables writing results, including 
simulated data, to file.
}
