% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pario.R
\name{read_pars}
\alias{read_pars}
\title{Read the parameters written to file}
\usage{
read_pars(filename)
}
\arguments{
\item{filename}{(character)}
}
\value{
list with elements:
\itemize{
     \item{$meta}{list of key value pairs of meta data}
     \item{$pars}{labeled matrix of parameters}
}
}
\description{
Read the parameters written to file
}
