% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_plink_snps}
\alias{sample_plink_snps}
\title{Store model parameters from a bpar file.}
\usage{
sample_plink_snps(n_snps, in_prefix, outdir)
}
\arguments{
\item{n_snps}{(int)
The desired number of variants to sample.}

\item{in_prefix}{(string)
The path and base filename (without extension) of the Plink dataset 
from which to sample snps.}

\item{outdir}{(string)
The directory path in which to save a Plink-formatted file of sampled 
variants.}
}
\value{
A list with (1) the path to the sampled variants file and (2) the 
     vector of sampled variants.
}
\description{
Produce one sample of n snps randomly sampled from a plink bim file.
}
