% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_snps}
\alias{sample_snps}
\title{Produce one or multiple sets of SNPs randomly sampled
from an input SNP set}
\usage{
sample_snps(
  input_snps,
  keep = 50000,
  iterations = 1,
  save_file = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{input_snps}{(character)
A vector of SNP variant names from which to sample. Names must be
in format 'chromosome:position', e.g. 1:2456462}

\item{keep}{(int)
(default 50000) The number of SNPs to keep}

\item{iterations}{(int)
(default 1) The number of random samples to produce}

\item{save_file}{(bool)
(default FALSE) Whether or not to save the SNP sample(s) to file(s)}

\item{output_dir}{(character)
The output directory in which to save the SNP sample(s)}
}
\value{
A list of length 'iterations' with multiple random SNP samples
}
\description{
Produce one or multiple sets of SNPs randomly sampled
from an input SNP set
}
