% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_snps_from_plink_files}
\alias{sample_snps_from_plink_files}
\title{Produce one or multiple sets of SNPs randomly sampled from an input
SNP set read in from Plink files, then use the new sample(s) to produce
a new Plink dataset(s)}
\usage{
sample_snps_from_plink_files(
  input_genotypes,
  snp_directory,
  output_dir,
  n_samples = 1,
  keep_files = TRUE
)
}
\arguments{
\item{input_genotypes}{(character)
The path/prefix of Plink files from which to sample SNPs}

\item{snp_directory}{(character)
The directory storing 1 or more text files listing sampled SNPS,
as produced by sample_snps()}

\item{output_dir}{(character)
The output directory in which to save the new Plink dataset(s)}

\item{n_samples}{(int)
(default 1) The number of random samples and Plink datasets 
to produce}

\item{keep_files}{(bool)
(default TRUE) Whether or not to keep Plink genotype files.
Use FALSE if only SNP files are desired}
}
\value{
A list of length n_samples of genotype datasets as produced
         by make_plink_dataset()
}
\description{
Produce one or multiple sets of SNPs randomly sampled from an input
SNP set read in from Plink files, then use the new sample(s) to produce
a new Plink dataset(s)
}
