% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_cv_results}
\alias{save_cv_results}
\title{Save the results from one k-fold cross-validation to a csv.}
\usage{
save_cv_results(cv_results, output_dir)
}
\arguments{
\item{cv_results}{(list)
     A list with results from one k-fold cross-validations, as output by 
     kfold_cv(). 
#' @param output_dir (character)
     The directory in which to save results}
}
\value{
None. Results are saved to file
}
\description{
Save the results from one k-fold cross-validation to a csv.
}
