% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_training_data}
\alias{save_training_data}
\title{Save training variant IDs and phenotype data from a .bpar file.}
\usage{
save_training_data(bpar_file)
}
\arguments{
\item{bpar_file}{(character)
Path to a model parameter file from which to extract metadata.}
}
\value{
None. Saves a text file of variant IDs and a csv file of training rfids 
and trait data into the same directory as the parameter file.
}
\description{
Reads a model parameter (.bpar) file to extract and save copies of phenotype 
data and variant IDs used to train the input RATTACA model.
}
