% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{simulator}
\alias{simulator}
\title{Make a closure of the LMM}
\usage{
simulator(intercept, u, int_se = NULL, u_se = NULL)
}
\arguments{
\item{intercept}{(double)}

\item{u}{((q markers,) vector) of BLUPs, that is the 
expected value of the random marker effect sizes}

\item{int_se}{(NULL | double) standard deviation of the LMM
random error term.}

\item{u_se}{(NULL | (q_markers,) vector) of stardard
error of the BLUPs}
}
\value{
(function) for generating samples from genotypes
     under the rrBLUP LMM
}
\description{
S
}
