% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{test_power_multi}
\alias{test_power_multi}
\title{Perform replicated power analyses on a set of hypothetical assignment groups.}
\usage{
test_power_multi(
  trait,
  genotypes,
  predictions,
  fitted_mod,
  group_size = c(0.01, 0.025, 0.05, 0.1),
  low_ids = NULL,
  high_ids = NULL,
  alpha = 0.05,
  reps = 10,
  tests_per_rep = 100,
  outdir = NULL
)
}
\arguments{
\item{trait}{(character)
The trait name}

\item{genotypes}{(matrix)
A named genotype matrix for all predicted animals}

\item{predictions}{(numeric)
The named vector of trait predictions that will be used to separate
animals into respective high or low assignment groups.}

\item{fitted_mod}{(list)
A fitted model object, as output by rattaca::fit()}

\item{group_size}{(numeric)
(default c(0.05, 0.1, 0.25, 0.33)) A numeric vector reflecting the 
sample size(s) desired for each assigned group to compare in a power 
analysis. Can be input either as proportions of the total predicted 
sample or sample counts. For example, 0.1 will compare the highest 10% 
and lowest 10% of predictions, and 20 will compare the highest 20 and 
lowest 20 predictions.}

\item{low_ids}{(character)
(default NULL) A vector of RFIDs that have already been assigned to a
'low' group based on trait predictions.}

\item{high_ids}{character)
(default NULL) A vector of RFIDs that have already been assigned to a
'high' group based on trait predictions.}

\item{alpha}{(numeric)
(default 0.05) The desired alpha value(s) (significance cutoff(s)) to 
use when determining statistical power to compare assigned groups.}

\item{reps}{(numeric)
(default 10) The desired number of replicated power analyses to conduct}

\item{tests_per_rep}{(numeric)
(default 100) The desired number of replicate simulations to conduct
for each power analysis. The number of comparisons to make between 
samples (per replicate).}

\item{outdir}{(character)
(default NULL) The desired output directory in which to save results}
}
\value{
A dataframe with one row per combination of desired group sample 
     size and alpha, and corresponding statistical power.
}
\description{
Conducts replicate runs of test_power() to calculate the mean and standard 
error of power analysis summary statistics.
}
