% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split}
\alias{train_test_split}
\title{Split a phenotypic dataset into train and test samples, and produce files
of RFIDs for either sample.}
\usage{
train_test_split(pheno_dat, trait, train_split = 0.7, train_dir, test_dir)
}
\arguments{
\item{pheno_dat}{(character or dataframe)
Either an R dataframe or the path to a csv file with phenotype data. 
Must have an 'rfid' column.}

\item{trait}{(character)
The name of the dataframe column (the trait name) on which to conduct
the train/test split}

\item{train_split}{(float)
(default 0.7) The desired frequency of samples to randomly include in 
the training set. The remainder (1 - train_split) will be included in 
the test set.}

\item{train_dir}{(character)
The directory path in which to store splitted training data.}

\item{test_dir}{(character)
The directory path in which to store splitted test data.}
}
\value{
A list with file paths to train and test data, file paths to text 
     files listing RFIDs present in either sample, and dataframes of 
     training and test data.
}
\description{
Split a phenotypic dataset into train and test samples, and produce files
of RFIDs for either sample.
}
