% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trait_groups}
\alias{trait_groups}
\title{Get assignment group designations (e.g., high vs low) for a set of 
     predictions.}
\usage{
trait_groups(preds, col, n_groups = 2)
}
\arguments{
\item{preds}{(dataframe)
A dataframe with at least one column of trait predictions.}

\item{col}{(character)
The name of the dataframe column on which to designate group assignments.}

\item{n_groups}{(int)
(default 2) The number of desired groups to split assignments into. The
default 2 splits into 'high' vs. 'low' groups, 3 splits into 
'low'/'mid'/'high', and other numbers split predictions into n_groups 
numeric quantiles.}
}
\value{
A vector of group assignments in the same order as col.
}
\description{
Get assignment group designations (e.g., high vs low) for a set of 
     predictions.
}
