% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{validate_test_preds}
\alias{validate_test_preds}
\title{Predict phenotypes in a test dataset with known trait values using 
parameters from a previously fit model, and assess model performance}
\usage{
validate_test_preds(phenotypes, genotypes, fitted_mod)
}
\arguments{
\item{phenotypes}{(numeric)
A named vector of observed phenotype values for a single trait}

\item{genotypes}{(matrix)
A matrix of m samples by n SNPs, with samples in the same
order as phenotypes}

\item{fitted_mod}{(list)
A fitted rattaca model, as output by fit()}
}
\value{
A list of (1) trait observations from the test set,
     (2) trait predictions for the test set, and (3,4,5)
     model performance metrics on the test set
}
\description{
Predict phenotypes in a test dataset with known trait values using 
parameters from a previously fit model, and assess model performance
}
