% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pario.R
\name{write_pars}
\alias{write_pars}
\title{Write inferred parameters to file}
\usage{
write_pars(trait, trait_file, genotype_prefix, pars, filename)
}
\arguments{
\item{trait}{(character) string specifying trait fit}

\item{trait_file}{(character) path to trait file used in fit}

\item{genotype_prefix}{(character) path to plink
genotype files used for fitting}

\item{pars}{(list) output from rattaca::fit}

\item{filename}{(character) path of file to write}
}
\description{
Write inferred parameters to file
}
