% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{zscore}
\alias{zscore}
\title{Standard-scale a dataset into Z-scores}
\usage{
zscore(data)
}
\arguments{
\item{data}{(numeric)
Any numeric vector}
}
\value{
A vector of Z-scores
}
\description{
Standard-scale a dataset into Z-scores
}
