% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argument}
\alias{argument}
\title{Specify command line argument values}
\usage{
argument(default_val = NULL, help = NULL, required = TRUE, type = "character")
}
\arguments{
\item{default_val}{(NULL | interger | double | string)
(default NULL)}

\item{help}{(string)
(default NULL) documentation to be printed with --help}

\item{required}{(bool) 
(default TRUE) whether the argument is required input}
}
\value{
(list) with correct key value pairs for parsing
}
\description{
Specify command line argument values
}
