% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{best_kfold_mod}
\alias{best_kfold_mod}
\title{Given results from multiple k-fold cross-validations, identify the
model with the best mean cross-validation performance.}
\usage{
best_kfold_mod(crossval_list)
}
\arguments{
\item{crossval_list}{(list)
A list with results from multiple k-fold cross-validations. Each
element must be a list as produced by kfold_cv() tested on
a different genomic dataset}
}
\value{
The integer value identifying the list element with the best
     mean cross-validation performance
}
\description{
Given results from multiple k-fold cross-validations, identify the
model with the best mean cross-validation performance.
}
