% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gen_sim_closure_r_sq}
\alias{gen_sim_closure_r_sq}
\title{Generate an LMM simulation closure from the r_sq}
\usage{
gen_sim_closure_r_sq(r_sq, genotypes, var_u = 1, intercept = 0)
}
\arguments{
\item{r_sq}{(float [0,1))
the coefficient of determination, a value 0 <= r_sq <  1
exclusive, between genetics and the simulated trait
measurements under the linear mixed model}

\item{genotypes}{((n_samples, q_markers) matrix)
genotype values}

\item{var_u}{(float (>0))
the variance of the random effects}

\item{intercept}{(float)
the intercept of the model}
}
\value{
(function)
}
\description{
Generate an LMM simulation closure from the r_sq
}
