% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_and_prepare_trait_data}
\alias{load_and_prepare_trait_data}
\title{load trait data from csv and set row index to given column}
\usage{
load_and_prepare_trait_data(filename, col_to_rownames, trait_name)
}
\arguments{
\item{filename}{(character) path to trait csv table}

\item{col_to_rownames}{(character) name of column whose
values will be the row index names}

\item{trait_name}{(character) name of column whose values
consists of normalized trait measurements}
}
\value{
(dataframe)
}
\description{
load trait data from csv and set row index to given column
}
