% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_composite_trait}
\alias{make_composite_trait}
\title{Produce a "composite" trait from the predictions of multiple individual 
traits, to be used in downstream assignment.}
\usage{
make_composite_trait(
  preds,
  traits,
  negate,
  invert,
  metric = c("rank", "zscore"),
  stat = c("mean", "median"),
  new_trait,
  plot_mar = c(0, 0, 8, 0),
  plot_oma = c(0, 0, 2, 2),
  title_line = 7,
  output_dir = NULL
)
}
\arguments{
\item{preds}{(dataframe or character)
A dataframe or csv path to a file containing merged RATTACA predictions 
for multiple traits.}

\item{traits}{(character)
Vector of trait names to be included in calculating the composite trait.}

\item{negate}{(logical)
A logical vector denoting, in the order of 'traits', whether to negate 
the predicted values for each trait. Negating multiplies each prediction 
by -1, effectively turning low ranks (or z-scores) high and high ranks
low.}

\item{invert}{(logical)
A logical vector denoting, in the order of 'traits', whether to invert 
the predicted values for each trait. Inverting divides 1/prediction, 
effectively turning intermediate z-scores extreme and extreme z-scores 
intermediate, or large ranks small and small ranks large.}

\item{metric}{(string)
Which prediction metric to use (either 'rank' or 'zscore') in 
calculating the composite trait.}

\item{stat}{(string)
Which summary statistic to use (either 'mean' or 'median') in 
calculating the composite trait. That is, the composite will be 
calculated as either the mean or median of all individual traits 
(per rat) after negation and/or inversion (or neither).}

\item{new_trait}{(string)
The name of the new composite trait}

\item{plot_mar}{(numeric)
(Defualt c(0,0,8,0)) A vector of plotting margins used to plot trait 
correlations This may need tweaking for aesthetic plotting depending on 
trait and variable names.}

\item{plot_oma}{(numeric)
(Default c(0,0,2,2)) A vector of outer plotting margins used to plot 
trait correlations This may need tweaking for aesthetic plotting 
depending on trait and variable names.}

\item{title_line}{(numeric)
(Default 7) The line above the correlation plot on which to print the
plot title. This may need tweaking for aesthetic plotting.}

\item{output_dir}{(string)
(Default NULL) The directory path in which to save output files}
}
\value{
Returns nothing to the R console. Saves csv files for trait 
     correlations and composite trait values, plus png files plotting trait
     correlations.
}
\description{
Reads trait predictions and modifies them to negate or invert their ranks 
or z-scores, then averages traits per individual to produce the composite. 
This allows prioritizing multiple traits at once for assignment, and trait 
modifications allow prioritization of rats that are, say, intermediate for 
one trait or whose set of traits are most consistently extreme, whether 
positively or negatively correlated as desired.
}
