% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict_lmm}
\alias{predict_lmm}
\title{Predict trait values from genotype}
\usage{
predict_lmm(genotypes, u, intercept)
}
\arguments{
\item{genotypes}{((n sample, q markers) array | matrix) of
sample genotypes.  Genotypes should be encoded 1,0,-1}

\item{u}{((q markers,) vector | (q markers,1) matrix) of
effect sizes for random effects}

\item{intercept}{(double)}
}
\value{
predictions
}
\description{
The `predict_lmm` function returns the value of a 
quantitative trait from an animal's genotype and
provided parameters inferred from a reference data set.
}
