% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_preds}
\alias{summarize_preds}
\title{Produce a file summarizing predictions model performance.}
\usage{
summarize_preds(
  traits,
  results_dir,
  basename = "all_traits",
  pheno_dict = NULL
)
}
\arguments{
\item{traits}{(character)
A vector of trait names(s) (corresponding to directory names with 
prediction results) to incorporate into the summary}

\item{results_dir}{(string)
The directory path to the results folder housing output directories 
named per trait. The function will search for files within each of 
the directories named in 'traits'.}

\item{basename}{(string)
The base name for the output summary file. The final output will be 
named <basename>_summary_<datestamp>.csv.}

\item{pheno_dict}{(string)
(Default NULL) The path to a phenotype data dictionary csv. The file 
must have columns 'trait', 'description', 'covariates', 'heritability', 
'variable_used', and 'source_file'.}
}
\value{
A list with model parameters, BLUPs, and performance statistics, 
     as output by fit().
}
\description{
Reads prediction results and a csv of trait metadata from multiple trait-
specific directories to merge metadata with model performance statistics.
}
